pyproject.toml
[tool.poetry]
name = "attendance-api"
version = "0.1.0"
description = "A Python-based microservice for handling attendance records"
authors = ["Opstree Solutions"]
license = "Apache"
readme = "README.md"
packages = [
    {include = "client"},
    {include = "models"},
    {include = "router"},
    {include = "utils"}
]

[tool.poetry.dependencies]
python = "^3.11"
peewee = "^3.16.2"
voluptuous = "^0.13.1"
redis = "^4.6.0"
json-logging-py = "^0.2"
jsonformatter = "^0.3.2"
python-json-logger = "^2.0.7"
flask-caching = "^2.0.2"
prometheus-flask-exporter = "^0.22.4"
flasgger = "0.9.7b2"
pytest-mock = "^3.11.1"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
Flask = "^2.3.2"
PyYAML = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = [
    "app.py"
]
